+++ .github/workflows/run.yml
@@ -93,7 +93,7 @@ jobs:
       - name: Generate Google API Token
         shell: bash
         run: |
-          poetry run python services/oauth_credentials/config_creds.py --save_dir .
+          poetry run python services/google_integration/config_creds.py --save_dir .
         working-directory: ${{ github.workspace }}
         env:
           GOOGLE_ACCESS_TOKEN: ${{ secrets.GOOGLE_ACCESS_TOKEN }}
