--- .github/workflows/setup.yml
@@ -1,65 +1,0 @@
-name: One-Time Setup
-
-on:
-  workflow_dispatch:
-
-jobs:
-  setup:
-    runs-on: ubuntu-latest
-
-    steps:
-      - name: Set up Node.js environment
-        uses: actions/setup-node@v3
-        with:
-          node-version: 16
-
-      - name: Create Labels
-        run: |
-          echo "Setting up labels..."
-          curl -X POST \
-            -H "Authorization: Bearer ${{ secrets.PERSONAL_ACCESS_TOKEN }}" \
-            -H "Accept: application/vnd.github.v3+json" \
-            https://api.github.com/repos/${{ github.repository }}/labels \
-            -d '{"name": "major", "color": "ff0000", "description": "Major version bump"}'
-      
-          curl -X POST \
-            -H "Authorization: Bearer ${{ secrets.GITHUB_TOKEN }}" \
-            -H "Accept: application/vnd.github.v3+json" \
-            https://api.github.com/repos/${{ github.repository }}/labels \
-            -d '{"name": "minor", "color": "ffa500", "description": "Minor version bump"}'
-      
-          curl -X POST \
-            -H "Authorization: Bearer ${{ secrets.GITHUB_TOKEN }}" \
-            -H "Accept: application/vnd.github.v3+json" \
-            https://api.github.com/repos/${{ github.repository }}/labels \
-            -d '{"name": "patch", "color": "00ff00", "description": "Patch version bump"}'
-
-
-      - name: Configure Branch Protection
-        run: |
-          echo "Setting up branch protection..."
-          curl -X PUT \
-            -H "Authorization: Bearer ${{ secrets.PERSONAL_ACCESS_TOKEN }}" \
-            -H "Accept: application/vnd.github.luke-cage-preview+json" \
-            https://api.github.com/repos/${{ github.repository }}/branches/main/protection \
-            -d '{
-              "required_status_checks": {
-                "strict": true,
-                "contexts": []
-              },
-              "enforce_admins": false,
-              "required_pull_request_reviews": {
-                "required_approving_review_count": 1,
-                "dismiss_stale_reviews": true,
-                "require_code_owner_reviews": false,
-                "require_last_push_approval": false
-              },
-              "restrictions": null,
-              "required_signatures": true,
-              "required_conversation_resolution": true,
-              "required_linear_history": false,
-              "required_deployments": [],
-              "allow_force_pushes": false,
-              "allow_deletions": false
-            }'
-
