+++ services/google_integration/gmail_services.py
@@ -7,7 +7,7 @@ from bs4 import BeautifulSoup
 from dateutil import parser
 from googleapiclient.discovery import build
 from googleapiclient.errors import HttpError
-from oauth_credentials.authentification import (
+from services.google_integration.authentification import (
     load_credentials,
     print_token_ttl,
     refresh_access_token,
