--- services/oauth_credentials/__init__.py
@@ -1,1 +1,0 @@
-# This file is intentionally left empty.
